fashion-mnist-dl/requirements.txt
# Core
tensorflow>=2.12.0
numpy>=1.23.0
pandas>=1.5.0
matplotlib>=3.6.0
seaborn>=0.12.0
scikit-learn>=1.2.0
Pillow>=9.4.0
pyyaml>=6.0

# Profiling & notebooks
ydata-profiling>=4.5.0
ipywidgets>=8.0.0

# Dev / testing
pytest>=7.0
black>=23.0
flake8>=6.0
